:art: extract reusable code from "SettingForm" to "FormInputGroup"
